refactor(ResultPage): move component to routes/shared/components
