+++ app/routes/beratungshilfe.vorabcheck.ergebnis.$.tsx
@@ -1,2 +1,2 @@
 export { loader } from "~/routes/shared/result.server";
-export { ResultPage as default } from "../components/resultPage/ResultPage";
+export { ResultPage as default } from "./shared/components/ResultPage";
